requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.58.1
python-dotenv==1.0.0
reportlab==4.0.7
PyPDF2==3.0.1
